fix issue with cancelling/deleting jobs

+++ app/routes_jobs.go
@@ -42,27 +42,42 @@ func (a *Application) JobsCreate(c *gin.Context, job string) {
 }
 
 func (a *Application) JobsDelete(c *gin.Context, id string, hard bool) {
-	filter := bson.M{"_id": id}
 	if id == string(minion.StatusPending) && !hard {
-		filter = bson.M{"status": minion.StatusPending}
-	}
-	if id == string(minion.StatusFailed) && hard {
-		filter = bson.M{"status": minion.StatusFailed}
-	}
-	if id == string(minion.StatusCancelled) && hard {
-		filter = bson.M{"status": minion.StatusCancelled}
-	}
-
-	if hard {
-		if _, err := app.DB.Minion.Collection.DeleteMany(context.Background(), filter); err != nil {
-			c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-			return
-		}
-	} else {
+		filter := bson.M{"status": minion.StatusPending}
 		if _, err := app.DB.Minion.Collection.UpdateMany(context.Background(), filter, bson.M{"$set": bson.M{"status": minion.StatusCancelled}}); err != nil {
 			c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 			return
 		}
+		c.JSON(http.StatusOK, gin.H{"error": false})
+		return
+	} else if id == string(minion.StatusFailed) && hard {
+		filter := bson.M{"status": minion.StatusFailed}
+		if _, err := app.DB.Minion.Collection.DeleteMany(context.Background(), filter); err != nil {
+			c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+			return
+		}
+		c.JSON(http.StatusOK, gin.H{"error": false})
+		return
+	} else if id == string(minion.StatusCancelled) && hard {
+		filter := bson.M{"status": minion.StatusCancelled}
+		if _, err := app.DB.Minion.Collection.DeleteMany(context.Background(), filter); err != nil {
+			c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+			return
+		}
+		c.JSON(http.StatusOK, gin.H{"error": false})
+		return
+	}
+
+	j, err := app.DB.Minion.Get(id, &Minion{})
+	if err != nil {
+		c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	j.Status = string(minion.StatusCancelled)
+	if err := app.DB.Minion.Save(j); err != nil {
+		c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
 	}
 
 	c.JSON(http.StatusOK, gin.H{"error": false})
